Couchbase - Rebalancing nodes - payload refactoring. Signed off by Tusa Mihai: tusa.mihai@gmail.com
