+++ cs-couchbase/src/main/java/io/cloudslang/content/couchbase/actions/cluster/RebalancingNodes.java
@@ -7,7 +7,7 @@
  * http://www.apache.org/licenses/LICENSE-2.0
  *
  *******************************************************************************/
-package io.cloudslang.content.couchbase.actions.nodes;
+package io.cloudslang.content.couchbase.actions.cluster;
 
 import com.hp.oo.sdk.content.annotations.Action;
 import com.hp.oo.sdk.content.annotations.Output;
